Improved empty block handling
